+++ lib/animated_to.dart
@@ -1,4 +1,4 @@
 library animated_to;
 
-export 'src/animated_to_container.dart';
+export 'src/animated_to_boundary.dart';
 export 'src/widget.dart';