+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Fixed tiny lint issues.
 
 ## [0.6.0] - 2025.12.5
-- Add `AnimatedToContainer` to detect gesture to descendant `AnimatedTo`s during their animation. 
+- Add `AnimatedToBoundary` to detect gesture to descendant `AnimatedTo`s during their animation. 
 
 ## [0.5.0] - 2025.12.3
 - `animated_to` now depends on `motor` instead of `springster`. Thank you [@timcreatedit](https://github.com/timcreatedit)